Fix + ADDED AFK RESTARTER

+++ system/plugins/afk.py
@@ -2,9 +2,10 @@
 
 
 import asyncio
-from pyrogram.raw.types.message import Message
-from pyrogram.types import update
-from system.datas_sqlite.afk_sqlite import get_afk, get_reason, update_afk, del_afk
+import logging
+import re
+from pyrogram.types import update, Message
+from system.datas_sqlite.afk_sqlite import ex, get_afk, get_reason, update_afk, del_afk, restart
 
 from system import COMMAND_HELP, OWNER, light, HNDLR
 from system.Config import Variable
@@ -15,30 +16,31 @@ from system import (
 from pyrogram import filters
 
 
-
-@light.on(["afk"])
+@light.on("afk")
 
 async def _(client, message ):
 
     try:
 
      if get_afk():
+         logging.info("Returned as afk enabled")
          return
     except IndexError:
         pass
 
-    try:
-        txt = message.text
-        reason = txt.split()[1:]
-        reason=" ".join(reason)
-    except IndexError:
-        reason = "**Contact me when i'm back alive, till [AFK]**\n __This is an automated message__"
+    txt = message.text
+    if " " in txt:
         
-    update_afk("True", reason)
+     reason = txt.split()[1:]
+     reason=" ".join(reason)
+    else:
+     reason = "**Contact me when i'm back alive, till [AFK]**\n __This is an automated message__"
+    await message.edit(f"**Afk update from now onwards i'll handle any kind of updates**\n\n**Reason** - {reason}") and     update_afk("True", reason)
     
-    await message.edit(f"**Afk update from now onwards i'll handle any kind of updates**\n\n**Reason** - {reason}")
 
-@app.on_message(~filters.bot & (filters.mentioned | filters.private) & ~filters.user('keinshin'))
+
+
+@app.on_message(~filters.bot & filters.mentioned & filters.private)
 async def n(client, message):
   user = await app.get_me()
   usr = await app.get_users(user.id)
@@ -60,23 +62,34 @@ __I was Last online __ - {user.status}
 """ 
   else:
        msg = get_reason()
-     
-  await app.send_message(message.chat.id, text=msg, reply_to_message_id=message.message_id)
+  msg=await app.send_message(message.chat.id, text=msg, reply_to_message_id=message.message_id)
+
+  if not get_afk():
 
 
-@app.on_message(filters.user(OWNER))
-async def sed(client, message:Message):
-     
-    try:
+      await msg.delete()
 
-     del_afk()
-    
-     msg=await app.send_message(message.chat.id, "**AFK BACK NORMAL [ONLINE]**" )
-     await asyncio.sleep(1)
-     await app.delete_messages(msg.chat.id, msg.message_id, revoke=True)
-    except Exception:
-        pass
+@app.on_message(filters.me)
+async def sed(client, message: Message):
+      if message.text.startswith(".afk"):
+            return
 
+      try:   
+
+
+       if get_afk():
+         del_afk()
+        
+         msg=await app.send_message(message.chat.id, "**AFK BACK NORMAL [ONLINE]**" )
+         await asyncio.sleep(1)
+         await app.delete_messages(msg.chat.id, msg.message_id, revoke=True)
+      except Exception:
+         pass
+
+@light.on("rerun")
+async def du(client, message):
+  restart()
+  await message.edit("**Afk restarted**")
 COMMAND_HELP.update({
     "afk": f"`{HNDLR} afk` `(reason) or default`",
     "afk's help": "**USE**: __Creates an automated message by your bot when you are afk__"
